+++ src/App.js
@@ -20,5 +20,5 @@ class App {
 
 module.exports = App;
 
-// let app = new App();
-// app.play();
+let app = new App();
+app.play();
