+++ src/Printer.js
@@ -1,11 +1,12 @@
 const MissionUtils = require('@woowacourse/mission-utils');
-const PLACE = {
-	FIRST: 7,
-	SECOND: 6,
-	THIRD: 5,
-	FOURTH: 4,
-	FIFTH: 3,
-};
+const { PLACE } = require('./constants');
+// const PLACE = {
+// 	FIRST: 7,
+// 	SECOND: 6,
+// 	THIRD: 5,
+// 	FOURTH: 4,
+// 	FIFTH: 3,
+// };
 
 class Printer {
 	printLottoBundle(lottoBundle) {
