+++ src/Prize.js
@@ -1,27 +1,28 @@
 const MissionUtils = require('@woowacourse/mission-utils');
 const Printer = require('./Printer');
 const Validator = require('./Validator');
-const PLACE = {
-	FIRST: 7,
-	SECOND: 6,
-	THIRD: 5,
-	FOURTH: 4,
-	FIFTH: 3,
-};
+const { PLACE, TYPE, PRIZE } = require('./constants');
+// const PLACE = {
+// 	FIRST: 7,
+// 	SECOND: 6,
+// 	THIRD: 5,
+// 	FOURTH: 4,
+// 	FIFTH: 3,
+// };
 
-const TYPE = {
-	LOTTO: '로또',
-	WINNINGNUM: '당첨',
-	BONUS: '보너스',
-};
+// const TYPE = {
+// 	LOTTO: '로또',
+// 	WINNINGNUM: '당첨',
+// 	BONUS: '보너스',
+// };
 
-const PRIZE = {
-	FIRST: 2000000000,
-	SECOND: 30000000,
-	THIRD: 15000000,
-	FOURTH: 50000,
-	FIFTH: 5000,
-};
+// const PRIZE = {
+// 	FIRST: 2000000000,
+// 	SECOND: 30000000,
+// 	THIRD: 15000000,
+// 	FOURTH: 50000,
+// 	FIFTH: 5000,
+// };
 
 class Prize {
 	constructor() {
