+++ src/Validator.js
@@ -1,8 +1,9 @@
-const TYPE = {
-	LOTTO: '로또',
-	WINNINGNUM: '당첨',
-	BONUS: '보너스',
-};
+// const TYPE = {
+// 	LOTTO: '로또',
+// 	WINNINGNUM: '당첨',
+// 	BONUS: '보너스',
+// };
+const { TYPE } = require('./constants');
 const regex = /[^0-9]/;
 
 class Validator {
