+++ src/constants.js
@@ -1,0 +1,27 @@
+const PLACE = {
+	FIRST: 7,
+	SECOND: 6,
+	THIRD: 5,
+	FOURTH: 4,
+	FIFTH: 3,
+};
+
+const TYPE = {
+	LOTTO: '로또',
+	WINNINGNUM: '당첨',
+	BONUS: '보너스',
+};
+
+const PRIZE = {
+	FIRST: 2000000000,
+	SECOND: 30000000,
+	THIRD: 15000000,
+	FOURTH: 50000,
+	FIFTH: 5000,
+};
+
+module.exports = {
+	PLACE,
+	PRIZE,
+	TYPE
+};
